+++ src/test/java/org/geowe/service/geometry/test/AllGeometryTests.java
@@ -5,8 +5,8 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({ BufferTest.class, IntersectionTest.class,
-	UnionTest.class, CentroidTest.class, EnvelopeTest.class, DifferenceTest.class })
+@SuiteClasses({ BufferTest.class, IntersectionTest.class, UnionTest.class, CentroidTest.class, EnvelopeTest.class,
+		DifferenceTest.class, SymDifferenceTest.class })
 public class AllGeometryTests {
 
 }