+++ src/test/java/org/geowe/service/geometry/test/AllGeometryTests.java
@@ -6,7 +6,7 @@ import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
 @SuiteClasses({ BufferTest.class, IntersectionTest.class, UnionTest.class, CentroidTest.class, EnvelopeTest.class,
-		DifferenceTest.class, SymDifferenceTest.class })
+		DifferenceTest.class, SymDifferenceTest.class, IntersectTest.class })
 public class AllGeometryTests {
 
 }