+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
-You can see the previous changelog [here] (http://www.geowe.org/index.php?id=changelog)
+
 
 ## [Unreleased]
 ### Added